fix(attn): the YaRN angle, the -1 block-table sentinel, and a tile table that named the wrong Bq (#1630 #1678 #1679) (#1704)
Three attention findings, one subsystem.

kept calling __sinf/__cosf on an unreduced argument, and the long-context
regression test could not see it: it runs at the default ext_factor = 0.0f, so
it takes the linear branch.

Both halves needed fixing. Reducing alone left it 1.3e-3 out at position 131071
against double truth, because the angle itself - position times frequency - does
not survive float there. theta_extrap is double at all four call sites now
(rope.cu x3, mtp_forward.cu), rope_yarn blends in double, and rope_sincos moved
into rope_yarn.cuh so both users share it.

New test drives the YaRN branch to 131071 against a double oracle. Removing the
reduction alone turns it red.

StreamingLLM writes -1; the FP16 kernel has skipped it since #963 as
defense-in-depth; FP8, FP8-tile, INT8, INT4, NVFP4 and NVFP4-TC read it
straight into a pointer. 12 guards, plus the tiled kernel, which prefetches
through a cp_async ring and cannot `continue`: it clamps the address into the
pool and drops the tokens with its validity mask instead.

Cost, 10 alternating runs, Qwen3-8B-Q8_0.gguf --kv-fp8, tg128:

    guard    median 384.88  mean 381.14  sd 4.1%
    noguard  median 396.41  mean 387.51  sd 6.4%
    paired: the guarded arm is FASTER in 4 of 10 rounds

Not separable from the noise: the spread of each arm exceeds the -2.9% gap
between their medians, and the sign flips round to round.

The test needed sharpening to be worth anything. With only -1 in the table it
passed against the unguarded kernel: that read lands one block before the pool,
which is still mapped, so the kernel returns quiet garbage rather than
faulting. A second, far-negative entry leaves the mapping, and then the
unguarded kernel fails with "an illegal memory access was encountered". One
guard covers both.

fmha_sm120_prefill's first three branches compare against max_smem/2, not
max_smem, so three of the comment's five rows named a Bq the code never picks.
Computed from compute_smem_sm120 against the measured
cudaDevAttrMaxSharedMemoryPerBlockOptin of 101376 (occ2_cap 50688):

    HD    Bkv  Bq  smem      comment said
    64     64  64  48.5 KB   128
    96     64  32  38.2 KB   64
    128    64  32  48.2 KB   64
    256    64  32  88.2 KB   32   (correct)
    512    32  16  82.1 KB   16   (correct)

Corrected in the kernel and in docs/internals/KERNELS.md, which carried the
same claim ("degrades to Bq=64").

Pins: two allowlist code_loc values moved by the guards, the unlaned GTest
count 973 -> 975 (both new tests need a GPU), and attention_paged_fp8.cu
crossed 600 by three lines and is allowlisted with that as its reason.

test-attention 215 passed, RoPE/MRope 14 passed, CPU lane 7/7.

+++ src/compute/mtp_forward.cu
@@ -371,8 +371,9 @@ __global__ void mtp_mrope_kernel(
         float c, s;
         if (ext_factor != 0.0f) {
             // YaRN mode: per-dimension frequency blending.
-            float theta_extrap =
-                static_cast<float>(pos) / powf(theta, static_cast<float>(2 * k) / static_cast<float>(rope_dim));
+            double theta_extrap = static_cast<double>(pos) /
+                                  pow(static_cast<double>(theta),
+                                      static_cast<double>(2 * k) / static_cast<double>(rope_dim));
             rope_yarn(theta_extrap, inv_scaling, corr_dim_0, corr_dim_1, 2 * k, ext_factor, attn_factor, c, s);
         } else {
             // Linear mode: frequency = theta^(-2k/rope_dim), scaled by inv_scaling.
